+++ FMI3_to_VDM/src/main/java/fmi2vdm/elements/ModelAttributes.java
@@ -29,6 +29,8 @@
 
 package fmi2vdm.elements;
 
+import java.math.BigInteger;
+
 import org.xml.sax.Attributes;
 import org.xml.sax.Locator;
 
@@ -45,7 +47,7 @@ public class ModelAttributes extends Element
 	private String generationTool;
 	private String generationDateAndTime;
 	private String variableNamingConvention;
-	private Long numberOfEventIndicators;
+	private BigInteger numberOfEventIndicators;
 
 	public ModelAttributes(Attributes attributes, Locator locator)
 	{
